eventpoi/requirements.txt
Django>=1.6
South>=0.8.1
django-sortedm2m>=0.6.1
django-model-utils>=2.0.3
psycopg2>=2.5.1
django-photologue>=2.8
django-tastypie>=0.10.0
django-redis-cache>=0.13
django-registration>=1.0
django-crispy-forms>=1.4.0